requirements.txt
streamlit
ultralytics
opencv-python-headless
Pillow


